+++ src/types.ts
@@ -1,7 +1,12 @@
-// tslint:disable-next-line:ban-types
+// tslint:disable:ban-types
+
 export type Method = Function;
 
 export type Getter = () => any;
+
+export interface Constructor<T> {
+    new(...args: any[]): T;
+}
 
 export interface IMap<T> { 
     [key: string]: T;